requirements.txt
alive_progress==2.4.1
colorthief==0.2.1
numpy==1.22.4
Pillow==9.4.0
progress==1.6
python-dotenv==1.0.0
requests==2.28.1
requests_cache==0.9.8
